Separate TextareaInput into Input and Textarea
