+++ src/components/form/index.ts
@@ -1,0 +1,2 @@
+export { default as Textarea } from './textarea';
+export { default as Input } from './input';
