+++ src/components/input/index.ts
@@ -1,2 +1,1 @@
 export { default as Input } from './input';
-export { default as TextareaInput } from './textarea-input';
